Move all payment functions from core.view to core.payments.mock
